% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{test.points}
\alias{test.points}
\title{Random field of points}
\usage{
test.points(n = 100L, TwoD = FALSE)
}
\arguments{
\item{n}{integer [1]; number of points}

\item{TwoD}{logical [1];
should the points be distributed in only two dimensions?}
}
\value{
\link[data.table]{data.table} with columns:\cr
\code{..$x,y,z} (num): position (\code{z} only if \code{ThreeD})\cr
\code{..$m} (num): mass (or weight) of points
}
\description{
Test and explore the functionality of \code{\link{coalesce}} with a
 random field of weighted points generated by this function.
}
\examples{
test.points(5L)
test.points(150L, FALSE)
test.points(150L)
\dontrun{test.points(1e5)}

}
